Finally got thick lines working. I think they are too thick, even on a normal screen. Not sure why.  Also added a dashed line implementation to try.

+++ scenegraph/gridnode.h
@@ -23,7 +23,7 @@ public:
                 const QMatrix4x4 &ndcMatrix,
                 const QSize viewportSize,
                 const QColor &gridColor,
-                int lineWidth,
+                double lineWidth,
                 double eastingMin, double eastingMax,
                 double northingMin, double northingMax,
                 double gridSpacing);
